sstables: fix overflow in clustering key blocks header bit access
_ck_blocks_header is a 64-bit variable, so the mask should be 64 bits too.
Otherwise, a shift in the range 32-63 will produce wrong results.

Fix by using a 64-bit mask.

Found by Fedora 29's ubsan.

Fixes #3973.
Message-Id: <20181209120549.21371-1-avi@scylladb.com>

+++ sstables/index_entry.hh
@@ -203,11 +203,11 @@ private:
         } state = state::CLUSTERING_START;
 
         bool is_block_empty() {
-            return (ck_blocks_header & (1u << (2 * ck_blocks_header_offset))) != 0;
+            return (ck_blocks_header & (uint64_t(1) << (2 * ck_blocks_header_offset))) != 0;
         }
 
         bool is_block_null() {
-            return (ck_blocks_header & (1u << (2 * ck_blocks_header_offset + 1))) != 0;
+            return (ck_blocks_header & (uint64_t(1) << (2 * ck_blocks_header_offset + 1))) != 0;
         }
 
         bool no_more_ck_blocks() { return ck_range.empty(); }
